\encoding{UTF-8}
\name{ASSA-package}
\alias{ASSA-package}
\docType{package}
\title{Applied Singular Spectrum Analysis}
\description{
The package \pkg{ASSA} is an add-on tool for \R that implements time
series decomposition and modeling methods based on singular spectrum
analysis (SSA) and multivariate singular spectrum analysis (MSSA). The
current version of the package includes tools tailored for extracting
business cycles, and for computing trendlines.

For a complete list of functions, data sets and documentation, type
\code{help.start()} and follow the link to \pkg{ASSA} on the
Package Index.

Funding from the project INTERSTATA (Interdisciplinary Statistics in
Action), by the International Research and Partnership Fund, is
gratefully acknowledged.
}
\keyword{ASSA}